requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
attrs==25.1.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
cloudscraper==1.2.71
cssselect==1.2.0
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
fake-useragent==2.0.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
jiter==0.8.2
lxml==5.3.1
openai==1.65.1
outcome==1.3.0.post0
parse==1.20.2
pydantic==2.10.6
pydantic_core==2.27.2
pyee==11.1.1
pymongo==4.11.2
pyparsing==3.2.1
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==1.26.20
w3lib==2.3.1
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.21.0
